Update CircleCI configuration to use "sorotoki -i --approve" command
This commit updates the CircleCI configuration file to use the "sorotoki -i --approve" command instead of "clc;".
Update CircleCI configuration to use "sorotoki -i --approve" command

This commit updates the CircleCI configuration file to replace the "clc;" command with the "sorotoki -i --approve" command in the `matlab/run-command` job. This change ensures that the correct command is executed during the build process.