Refactor code to use `which` instead of `whereis`
- Replaced `whereis` with `which` in line 6
- Updated variable name from `installFile` to `installPath`
- Removed unnecessary substring operation on `installPath`

This commit refactors the code in the sorotoki.m file by replacing the deprecated function whereis with which. It also updates a variable name and removes an unnecessary substring operation.
Refactor code to use `which` instead of `whereis`

- Replaced `whereis` with `which` in line 6
- Updated variable name from `installFile` to `installPath`
- Removed unnecessary substring operation on `installPath`

This commit refactors the code in the sorotoki.m file by replacing the deprecated function whereis with which. It also updates a variable name and removes an unnecessary substring operation.