feat: Add pneunet_AMOLF mesh generation function
This commit adds the `pneunet_AMOLF` function to generate a mesh using the Mesh class. The function takes an optional argument for the number of elements and uses an image file located in the assets folder. The generated mesh is then simplified with a tolerance of 0.1.

The commit also includes the addition of the `Amolf_Actuator_25x118.png` image file to the assets folder, as well as updates to `.gitignore` and removal of unnecessary files (`test.m`) while adding a new file (`testfile.m`).